TheSparksFoundation GRIP_Tasks
Task 2 Problem
In this regression task we will predict the percentage of marks that a student is expected to score based upon the number of hours they studied. This is a simple linear number regression task as it involves just two numbers. What will be predicted score if a student study for 9.25 hours in a day?